Added Command class
To send and receive commands, Command class is added.
It has utilities to create executable, execute the command
and store output and error logs in specific files. This will help
to track output and errors generated because of execution of
command.

Also serializing and deserializing is of Command object is required
to send and receive object from 1 system to another over network.

Author: return007